requirements.txt
Flask==2.3.2
requests==2.31.0
sentence-transformers==2.2.2
rouge-score==0.1.2
scikit-learn==1.3.0
openai==1.3.0
numpy==1.24.3
boto3==1.28.44
python-dotenv==1.0.0 